requirements.txt
boto3==1.23.5
discord.py==1.7.3
matplotlib==3.4.3
pandas==1.3.4
python-dotenv==0.20.0
sagemaker==2.91.1
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.8.1
